requirements.txt
# Core dependencies
Flask>=3.0.0
FastAPI>=0.109.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0

# Database migrations
alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Code quality
black>=24.3.0
isort>=5.13.2
pre-commit>=3.6.0

# Additional utilities
python-multipart>=0.0.9  # for FastAPI form handling
Flask-Session>=0.6.0     # для сессий Flask

# Python-dotenv for environment variables
python-dotenv>=1.0.0
